lots of code duplication here, but implement PUT
My hope is to get a basic implementation of all the HTTP methods and
then start the refactoring work. Along the way, I'll have a bunch of
duplication and messy code, but it'll keep me moving forward. Perfect is
the enemy of good, and I don't want to fall into that too early :)